feat(web): refactor
